# Load and render parts of the gui at a later time