--- a/src/containers/costume-tab.dynamic.jsx
+++ b/src/containers/costume-tab.dynamic.jsx
@@ -1,0 +1,3 @@
+import {idleRequireWhileLoading} from '../lib/dynamic-render';
+
+export default idleRequireWhileLoading(() => require('./costume-tab.jsx'));