@mzgoddard 2019-04-08
- Fix lint