@paulkaplan 2019-05-23
Closing this for now, since we made progress with a smaller version. We may want to revisit this in the future as a model for further splitting up the load